refactor: 컴포넌트 분리하여 Suspense로 비동기 데이터 처리
